my_first_django/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==41.0.3
dateparser==1.1.8
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-admin-thumbnails==0.2.8
django-cacheops==7.0.1
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-jet-reboot==1.3.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-oauth-toolkit==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
easy-thumbnails==2.8.5
ecdsa==0.18.0
et-xmlfile==1.1.0
Faker==19.1.0
frozenlist==1.4.0
funcy==1.18
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.1
MarkupSafe==2.1.3
multidict==6.0.4
npm==0.1.1
numpy==1.25.2
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
optional-django==0.1.0
packaging==23.1
pandas==2.0.3
pika==1.3.2
Pillow==10.0.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
python-binance==1.0.17
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
